refactor(cli): move build module under commands
